+++ pythran/pythonic/include/__builtin__/file/read.hpp
@@ -1,9 +1,9 @@
 #ifndef PYTHONIC_INCLUDE_BUILTIN_FILE_READ_HPP
 #define PYTHONIC_INCLUDE_BUILTIN_FILE_READ_HPP
 
-#include "pythonic/utils/proxy.hpp"
-#include "pythonic/types/file.hpp"
-#include "pythonic/types/str.hpp"
+#include "pythonic/include/types/file.hpp"
+#include "pythonic/include/types/str.hpp"
+#include "pythonic/include/utils/proxy.hpp"
 
 namespace pythonic
 {
@@ -14,8 +14,8 @@ namespace pythonic
     namespace file
     {
 
-      types::str read(types::file &f, int size = -1);
-      types::str read(types::file &&f, int size = -1);
+      types::str read(types::file &f, long size = -1);
+      types::str read(types::file &&f, long size = -1);
 
       PROXY_DECL(pythonic::__builtin__::file, read);
     }
