+++ pythran/pythonic/include/__builtin__/file/readlines.hpp
@@ -1,10 +1,10 @@
 #ifndef PYTHONIC_INCLUDE_BUILTIN_FILE_READLINES_HPP
 #define PYTHONIC_INCLUDE_BUILTIN_FILE_READLINES_HPP
 
-#include "pythonic/utils/proxy.hpp"
-#include "pythonic/types/list.hpp"
-#include "pythonic/types/str.hpp"
-#include "pythonic/types/file.hpp"
+#include "pythonic/include/types/file.hpp"
+#include "pythonic/include/types/list.hpp"
+#include "pythonic/include/types/str.hpp"
+#include "pythonic/include/utils/proxy.hpp"
 
 namespace pythonic
 {
@@ -18,7 +18,7 @@ namespace pythonic
       template <class F>
       types::list<types::str> readlines(F &&f);
       template <class F>
-      types::list<types::str> readlines(F &&f, int sizehint);
+      types::list<types::str> readlines(F &&f, long sizehint);
 
       PROXY_DECL(pythonic::__builtin__::file, readlines);
     }
